fix(routes): wrap function in useMemo, rename components
