+++ src/AppRoutes.tsx
@@ -6,10 +6,10 @@ import { AuthView } from './modules/auth/views/AuthView';
 import { NoMatchView } from './modules/noMatch/views/NoMatchView';
 import { StudentAssignmentsView } from './modules/student/views/StudentAssignmentsView';
 import { TaskView } from './modules/student/views/TaskView';
-import { LessonView as StudentLessonView } from './modules/studentLessons/views/LessonView';
+import { StudentLessonView } from './modules/studentLessons/views/StudentLessonView';
 import { StudentLessons } from './modules/studentLessons/views/StudentLessons';
 import { TeacherAssignmentsView } from './modules/teacher/views/TeacherAssignmentsView';
-import { LessonView as TeacherLessonView } from './modules/teacherLessons/views/LessonView';
+import { TeacherLessonView } from './modules/teacherLessons/views/TeacherLessonView';
 import { TeacherLessons } from './modules/teacherLessons/views/TeacherLessons';
 import { useAppSelector } from './store';
 import { selectIsLoggedIn, selectUserType } from './store/slices/auth';
